% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-descriptions.R
\docType{data}
\name{k4_dsd}
\alias{k4_dsd}
\title{Four factor definitive screening design}
\format{
A matrix with 4-columns and 9-rows
}
\usage{
k4_dsd
}
\description{
Four factor definitive screening design
}
\keyword{datasets}
