% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-descriptions.R
\docType{data}
\name{k5_dsd}
\alias{k5_dsd}
\title{Five factor definitive screening design}
\format{
A matrix with 5-columns and 11-rows
}
\usage{
k5_dsd
}
\description{
Five factor definitive screening design
}
\keyword{datasets}
